requirements.txt
omegaconf
hydra-core
scipy
h5py
python-dateutil
tqdm
pandas
mne
psutil
tensorboardX
scikit-learn
tensorboard
torch_optimizer
tables
#git+https://github.com/ildoonet/pytorch-gradual-warmup-lr.git
